Move MediaStreamTrack API into separate module
